VR Reticle Gaze Target Test
Add a scene into Test Scenes called VRReticleGazeTarget. Add the
CardboardMain (Camera) to that Scene, removed the default Camera. Add
the Prefab CardboardReticle to the CardboardMain Head, (made it red for
clarity). Add the Raytrace Component to the Cardboard Main Camera. Add
the EventSystem UI to the Hierarchy, and attached the Gaze Input Module
Component and moved it above the Standalone Input Module to avoid
conflict. Made a SchoolOfFish Layer and assigned the Tuna to it. Set
the Raytrace on the Main Camera to the SchoolOfFish Layer. Pressing
Play: You can see the Reticle expand and contract as you hold “alt” and
pan over the fish and off the fish. This will be the targeting feature
for catching fish and other interactions.