targetTime and Debug for making fish disappear after 3 seconds.
In the VRReticleGazeTarget scene, and modifying the Cardboard asset
files: Added a timer and did some debug testing on the
CardboardReticle.cs file. There is a small bug where the reticle
doesn’t scale down when the fish disappears under it. But it does if it
enters and exits another fish.