Add VisitReef script for repeatedly going through a tunnel and arriving at a reef
Still a work-in-progress. The TransitTest scene shows current
behaviour. It stays at a reef for a couple seconds and then starts
moving towards a newly instantiated reef through a tunnel. The tunnels
are straight tubes created programatically. They're created parallel
to the ground plane in the direction you're looking when the time runs
out at the reef.

The tunnel and movement behaviour is in a prefab. Hoping to replace it
with more a curved tunnels, smoother movement, and animating
the fish ahead of you in the tunnel.

The reef created after each tunnel is a single prefab. Need to replace
that with a list of prefabs or similar.