Coral2D package made and deployed
In order to keep organized, I packaged the 2D coral, which makes it
easier to give to other people and load in any Unity project. It’s self
contained and updating is just releasing another package. … its just a
thought since some of the assets are getting further and further apart
from each other in the project.