VisitReefs: Allow components in reef prefabs to delay transit
Any component in the reef prefab can implement IReefVisitor and will
be passed a ReefStage object. The tunnel will not be created until all
visitors have called the finishStage method of their ReefStage object.

ReefWait is an example component that waits a set amount of time which
isn't very exciting.