FishSchool: Use localPosition and localRotation when updating fish transform and storing in octtree
The position and rotation properties are in world coordinates which
require walking up the hierarchy to calculate, and there's always
going to be at least one parent for the fishes.

Still use world position when comparing against lures or boundaries,
as those objects have a different transform parent.

Note that if the fish container transform is changed, then the octtree
gizmos are not drawn in the correct spot.